+++ R/remove_domain.R
@@ -1,0 +1,16 @@
+#' function to remove domain from path
+#' @param x path aka URL from which to first infer domain and then remove it
+remove_domain <- function(x){
+  unlist(lapply(
+    x,
+    function(x){
+      if( is.na(x) ){
+        return(x)
+      }else{
+        stringr::str_replace(x, paste0("^.*", guess_domain(x)), "")
+      }
+    }
+  ))
+}
+
+