+++ R/rt_cache.R
@@ -1,0 +1,3 @@
+#' get_robotstxt() cache
+rt_cache <- new.env( parent = emptyenv() )
+