+++ R/rt_get_comments.R
@@ -1,0 +1,11 @@
+
+#' extrcting comments from robots.txt
+#' @param txt content of the robots.txt file
+#' @keywords internal
+rt_get_comments <- function(txt){
+  txt      <- unlist(stringr::str_split(txt, "\n"))
+  clines   <- grep("#", txt)
+  ccontent <- stringr::str_extract(txt[clines], "#.*")
+  data.frame(line=clines, comment=ccontent, stringsAsFactors = FALSE)
+}
+